requirements.txt
python-dotenv
elevenlabs
googlemaps
cohere
